Main / Board 분리
